도메인 수정

+++ src/main/java/kr/hhplus/be/server/dto/point/RequestPointCharge.java
@@ -2,5 +2,5 @@ package kr.hhplus.be.server.dto.point;
 
 public record RequestPointCharge(
         long userId,
-        int userPoint
+        long userPoint
 ){}
